pyproject.toml
[tool.poetry]
name = "property"
version = "0.1.0"
description = ""
authors = ["Vijay"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
django-formtools = "^2.5.1"
django-debug-toolbar = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
